refactor(domain): VoiceRecordRepository 프로토콜 분리 및 폴더 구조 정리
- VoiceRecordRepository.swift 파일을 Repositories/VoiceRecords/ 폴더로 이동 및 기능별 프로토콜로 분리
- 각 프로토콜을 독립된 파일로 생성하여 유지보수성 향상
